Fix: rename main script file to match gem name
